# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WorldAlert API - Global alert aggregation
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AlertCategory = typeof AlertCategory[keyof typeof AlertCategory];


export const AlertCategory = {
  earthquake: 'earthquake',
  disaster: 'disaster',
  humanitarian: 'humanitarian',
  weather: 'weather',
  conflict: 'conflict',
  health: 'health',
  other: 'other',
} as const;

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Alert {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  category: AlertCategory;
  severity: AlertSeverity;
  lat: number;
  lng: number;
  /** @nullable */
  country: string | null;
  /** @nullable */
  region?: string | null;
  source: string;
  sourceUrl: string;
  publishedAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  magnitude?: number | null;
  /** @nullable */
  affectedPopulation?: number | null;
}

export type AlertStatsBySeverity = {[key: string]: number};

export type AlertStatsByCategory = {[key: string]: number};

export interface AlertStats {
  total: number;
  critical?: number;
  bySeverity: AlertStatsBySeverity;
  byCategory: AlertStatsByCategory;
  lastUpdated: string;
}

export type ListAlertsParams = {
category?: string;
severity?: string;
limit?: number;
};