# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WorldAlert API - Global alert aggregation
 * OpenAPI spec version: 0.1.0
 */
import type { AlertCategory } from './alertCategory';
import type { AlertSeverity } from './alertSeverity';

export interface Alert {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  category: AlertCategory;
  severity: AlertSeverity;
  lat: number;
  lng: number;
  /** @nullable */
  country: string | null;
  /** @nullable */
  region?: string | null;
  source: string;
  sourceUrl: string;
  publishedAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  magnitude?: number | null;
  /** @nullable */
  affectedPopulation?: number | null;
}