# lib/api-zod/src/generated/types/alertStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WorldAlert API - Global alert aggregation
 * OpenAPI spec version: 0.1.0
 */
import type { AlertStatsByCategory } from './alertStatsByCategory';
import type { AlertStatsBySeverity } from './alertStatsBySeverity';

export interface AlertStats {
  total: number;
  critical?: number;
  bySeverity: AlertStatsBySeverity;
  byCategory: AlertStatsByCategory;
  lastUpdated: string;
}